requirements.txt
google-adk>=1.0.0
python-dotenv>=1.0.0
litellm
openai
